refactor: extract overview generation into module
