+++ index.html
@@ -1332,5 +1332,6 @@
   <script src="translations.js"></script>
   <script src="lz-string.min.js"></script>
   <script src="script.js"></script>
+  <script src="overview.js"></script>
 </body>
 </html>
